+++ packages/data/index.js
@@ -1,0 +1,1 @@
+// This is just for TS declarations to work